Refactor main script into package
